+++ node/src/main/scala/coop/rchain/node/runtime/APIServers.scala
@@ -36,7 +36,7 @@ object APIServers {
       blockReportAPI: BlockReportAPI[F],
       networkId: String,
       shardId: String,
-      minPhloPrice: Int
+      minPhloPrice: Long
   )(
       implicit
       blockStore: BlockStore[F],
